GRP_DEMO/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2==2.9.9
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0